+++ src/components/ids-data-grid/demos/virtual-scroll-infinite-scroll.ts
@@ -80,17 +80,17 @@ columns.push({
 
 dataGrid.columns = columns;
 
-const setData = async () => {
+const fetchData = async (startIndex = 0) => {
   const res = await fetch(url);
   const data = await res.json();
-  dataGrid.data = data;
+  return data.splice(startIndex, 100);
+};
+
+const setData = async () => {
+  dataGrid.data = await fetchData();
 };
 
 setData();
-
-dataGrid.addEventListener('selectionchanged', (e: Event) => {
-  console.info(`Selection Changed`, (<CustomEvent>e).detail);
-});
 
 dataGrid.addEventListener('scrollstart', async (e: Event) => {
   console.info(`Virtual Scroll reached start`, (<CustomEvent>e).detail);
@@ -98,4 +98,11 @@ dataGrid.addEventListener('scrollstart', async (e: Event) => {
 
 dataGrid.addEventListener('scrollend', async (e: Event) => {
   console.info(`Virtual Scroll reached end`, (<CustomEvent>e).detail);
+  const endIndex = (<CustomEvent>e).detail?.value || 0;
+  const MAX_ROW_INDEX = 299;
+
+  if (endIndex < MAX_ROW_INDEX) {
+    const moreData = await fetchData(endIndex + 1);
+    dataGrid.appendData(moreData);
+  }
 });